Move AGW backend plugin processing funcs into their own files
