+++ internal/kgateway/extensions2/plugins/backend/agentgateway_mcp.go
